refactor: remove DateEditCell
